requirements.txt
requests~=2.32.3
beautifulsoup4~=4.12.3
pymysql~=1.1.1
lxml~=5.3.0
websocket-client~=1.8.0
selenium~=4.28.1